requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
alabaster==0.7.12
    # via sphinx
asgiref==3.5.0
    # via django
babel==2.9.1
    # via sphinx
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.11
    # via requests
django==4.0.2
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-auth-ldap
    #   django-filter
    #   django-select2
    #   djangorestframework
django-appconf==1.0.5
    # via django-select2
django-auth-ldap==4.0.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-select2==7.9.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
docutils==0.17.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
et-xmlfile==1.1.0
    # via openpyxl
flake8==4.0.1
    # via -r requirements.in
huey==2.4.3
    # via -r requirements.in
idna==3.3
    # via requests
imagesize==1.3.0
    # via sphinx
importlib-metadata==4.10.1
    # via sphinx
jinja2==3.0.3
    # via sphinx
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via flake8
openpyxl==3.0.9
    # via -r requirements.in
packaging==21.3
    # via sphinx
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.2.8
    # via python-ldap
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pygments==2.11.2
    # via sphinx
pymemcache==3.5.0
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
python-ldap==3.4.0
    # via django-auth-ldap
python-magic==0.4.24
    # via -r requirements.in
pytz==2021.3
    # via
    #   babel
    #   djangorestframework
requests==2.27.1
    # via sphinx
segno==1.4.1
    # via -r requirements.in
six==1.16.0
    # via pymemcache
snowballstemmer==2.2.0
    # via sphinx
sphinx==4.4.0
    # via
    #   -r requirements.in
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.0.0
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-mermaid==0.7.1
    # via -r requirements.in
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.2
    # via django
urllib3==1.26.8
    # via requests
zipp==3.7.0
    # via importlib-metadata
